Implements report URL
If a `report_url` is set in ignore_list.yaml, it will display a
report URL at the end of the anteater job (if the job fails, e.g
if finds a string match, or positive VT API match).

Individual project report URLs can be set within the
project_exceptions file.

https://github.com/anteater/anteater/issues/38